arg parsing, minor name changes and fixes
