+++ code/config.ini
@@ -1,0 +1,11 @@
+[DEFAULT]
+algorithm = ray_tracing
+scene = scene.obj
+environment_map = map.hdr
+resolution = 400x300
+
+[ray_tracing]
+max_depth = 5 ; Params for ray tracing
+
+[photon_mapping]
+photon_count = 100000 ; Params for photon mapping
