--- code/requirements.txt
@@ -1,9 +1,0 @@
-glfw
-numpy
-pyrr
-PyOpenGL
-matplotlib
-flake8
-black
-autopep8
-flake8-max-function-length
